requirements.txt
pandas
yfinance
matplotlib
python-dotenv
sec-api
google-generativeai

